+++ nix/test-overlay.nix
@@ -61,7 +61,10 @@ with final.stdenv; let
     mkDerivation {
       inherit name;
 
+      src = self;
+
       phases = [
+        "unpackPhase"
         "buildPhase"
         "checkPhase"
       ];
@@ -84,14 +87,17 @@ with final.stdenv; let
         ln -s ${toggleterm} $out/.config/nvim/site/pack/packer/start/toggleterm.nvim
         ${optionalString withTelescope "ln -s ${telescope-nvim} $out/.config/nvim/site/pack/packer/start/telescope.nvim"}
         ln -s ${./..} $out/.config/nvim/site/pack/packer/start/${name}
+        cp -r tests $out
+        # FIXME: Generating a config does not seem to be working. For now, there is a config saved in the tests directory.
+        # haskell-language-server-wrapper generate-default-config > $out/tests/hls.json
       '';
 
       checkPhase = ''
         export NVIM_DATA_MINIMAL=$(realpath $out/.config/nvim)
         export HOME=$(realpath .)
-        cd ${./..}
-        # TODO: split test directories by environment
-        nvim --headless --noplugin -u ${../tests/minimal.lua} -c "PlenaryBustedDirectory ${../tests} {minimal_init = '${../tests/minimal.lua}'}"
+        export TEST_CWD=$(realpath $out/tests)
+        cd $out
+        nvim --headless --noplugin -u ${../tests/minimal.lua} -c "PlenaryBustedDirectory tests {minimal_init = 'tests/minimal.lua'}"
       '';
     };
 in {