+++ tests/lsp_spec.lua
@@ -1,5 +1,4 @@
 local ht = require('haskell-tools')
--- local stub = require('luassert.stub')
 
 describe('LSP client API', function()
   ht.setup {}
@@ -9,5 +8,13 @@ describe('LSP client API', function()
   it('Can spin up haskell-language-server for Cabal project.', function()
     --- TODO: Figure out how to add tests for this
     print('TODO')
+  end)
+  it('Can load haskell-language-server config', function()
+    local settings = ht.lsp.load_hls_settings(os.getenv('TEST_CWD'))
+    assert.not_same(ht.config.options.hls.default_settings, settings)
+  end)
+  it('Falls back to default haskell-language-server config if none is found', function()
+    local settings = ht.lsp.load_hls_settings(os.getenv('TEST_CWD'), { settings_file_pattern = 'bla.json' })
+    assert.same(ht.config.options.hls.default_settings, settings)
   end)
 end)