+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Set up LSP client without `nvim-lspconfig` (removes the dependency).
 ### Added
 - `HlsStart`, `HlsStop` and `HlsRestart` commands.
+- Dynamically load `haskell-language-server` settings JSON from project root, if available.
 
 ## [1.7.0] - 2023-01-27
 ### Fixed